feat : add ConvNextModel on v2
