+++ v2/model.py
@@ -1,5 +1,6 @@
 import torch.nn as nn
 import torch.nn.functional as F
+import torchvision
 
 
 class BaseModel(nn.Module):
@@ -66,4 +67,29 @@ class MyModel(nn.Module):
         1. 위에서 정의한 모델 아키텍쳐를 forward propagation 을 진행해주세요
         2. 결과로 나온 output 을 return 해주세요
         """
+        return x
+
+class ConvNextModel(nn.Module):
+    def __init__(self, num_classes):
+        super().__init__()
+
+
+        # Initializing a model (with random weights) from the convnext-tiny-224 style configuration
+        self.convnext = torchvision.models.convnext_tiny(weights='IMAGENET1K_V1')
+        self.convnext.classifier = nn.Sequential(
+            nn.LayerNorm((768,1,1,), eps=1e-06, elementwise_affine=True),
+            nn.Flatten(start_dim=1, end_dim=-1),
+            # nn.Linear(in_features=768, out_features=1024, bias=True),
+            #nn.Linear(in_features=1024, out_features=num_classes, bias=True)
+            nn.Linear(in_features=768, out_features=num_classes, bias=True)
+        )
+
+
+    def forward(self, x):
+        """
+        1. 위에서 정의한 모델 아키텍쳐를 forward propagation 을 진행해주세요
+        2. 결과로 나온 output 을 return 해주세요
+        """
+        x=self.convnext(x)
+        
         return x
